requirements.txt
asttokens==3.0.0
autocommand==2.2.2
backports.tarfile==1.2.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.57.0
imbalanced-learn==0.13.0
importlib-metadata==8.6.1
inflect==7.3.1
ipykernel==6.29.5
ipython==9.1.0
ipython-pygments-lexers==1.1.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
more-itertools==10.3.0
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip==25.0.1
platformdirs==4.3.7
prompt-toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==79.0.0
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
threadpoolctl==3.6.0
tomli==2.0.1
tornado==6.4.2
traitlets==5.14.3
typeguard==4.3.0
typing-extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
wheel==0.45.1
zipp==3.21.0
